# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Job Matching API
  description: REST API for managing candidates, job postings, and AI-powered matching.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
  /candidates:
    get:
      tags:
        - Candidates
      summary: List all candidates with pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (starts at 0)
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 10
          description: Number of candidates per page
      responses:
        '200':
          description: A paginated list of candidates
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CandidateResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
    post:
      tags:
        - Candidates
      summary: Register a new candidate
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  minLength: 1
                  format: email
                password:
                  type: string
                  minLength: 1
                skills:
                  type: array
                  items:
                    type: string
                experienceYears:
                  type: integer
                resume:
                  type: string
                  format: binary
      responses:
        '201':
          description: Candidate created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResponse'
  /candidates/{id}:
    get:
      tags:
        - Candidates
      summary: Get candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Candidate data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResponse'
    delete:
      tags:
        - Candidates
      summary: Delete candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Candidate deleted
        '404':
          description: Candidate not found
    put:
      tags:
        - Candidates
      summary: Update an existing candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  minLength: 1
                  format: email
                password:
                  type: string
                  minLength: 1
                skills:
                  type: array
                  items:
                    type: string
                experienceYears:
                  type: integer
                resume:
                  type: string
                  format: binary
      responses:
        '200':
          description: Candidate updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResponse'
...
components:
  schemas:
    CandidateResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        resumeText:
          type: string
...